Update and rename User Input (Sacnner class).java to User Input (Scanner class).java

+++ Java/User Input (Scanner class).java
@@ -6,7 +6,7 @@ public class Solution {
         Scanner scan = new Scanner(System.in);
         int i = scan.nextInt();
         // Write your code here.
-        float d = scan.nextFloat();
+        double d = scan.nextDouble();
         scan.nextLine();
         String s = scan.nextLine();
         scan.close();
